Migrate comonent_manager to app v2.
Removes convenience ConnectToService() from ComponentManager. It will
live elsewhere than on the main ComponentManager service since we want
to be able to have finer control over the environment and capabilities
passed to new applications.

No longer build the example apps, as the pattern for gaining access to
the ComponentManager isn't 100% clear yet.

Change-Id: I74b7c4f5d8f2f4c3262dccf6ddabd944c67c986a